+++ src/entity/index.ts
@@ -31,7 +31,9 @@ export * from './PostKeyword';
 export * from './CommentMention';
 export * from './ReputationEvent';
 export * from './Submission';
-export * from './notifications';
+export * from './notifications/Notification';
+export * from './notifications/NotificationAvatar';
+export * from './notifications/NotificationAttachment';
 export * from './Feature';
 export * from './UserAction';
 export * from './ContentImage';
