+++ src/entity/notifications/NotificationPreferencePost.ts
@@ -1,6 +1,6 @@
 import { ChildEntity, ManyToMany, PrimaryColumn } from 'typeorm';
 import { Post } from '../posts';
-import { NotificationPreferenceType } from '../../notifications';
+import { NotificationPreferenceType } from '../../notifications/common';
 
 @ChildEntity(NotificationPreferenceType.Post)
 export class NotificationPreferencePost {
