+++ src/workers/notifications/worker.ts
@@ -1,7 +1,8 @@
 import { DataSource } from 'typeorm';
 import { FastifyLoggerInstance } from 'fastify';
 import { Message } from '../worker';
-import { NotificationBaseContext, NotificationType } from '../../notifications';
+import { NotificationBaseContext } from '../../notifications';
+import { NotificationType } from '../../notifications/common';
 
 export type NotificationHandlerReturn = {
   type: NotificationType;
